Added a basic unit test for GraphDatabase.

+++ pygs/graphserver/graphdb.py
@@ -10,11 +10,11 @@ class GraphDatabase:
     
     def __init__(self, sqlite_filename, overwrite=False):
         if overwrite:
-            try:
+            if os.path.exists(sqlite_filename):
                 os.remove( sqlite_filename )
-            except OSError:
-                pass
-            
+        elif not os.path.exists(sqlite_filename):
+            overwrite = True # force an init of the tables
+                
         self.conn = sqlite3.connect(sqlite_filename)
         
         if overwrite:
@@ -146,11 +146,10 @@ class GraphDatabase:
         return g
         
 
-if __name__=='__main__':
-    
+def main():
     if len(argv) < 2:
         print "usage: python graphdb.py [vertex1, [vertex2]]"
-        exit()
+        return
     
     graphdb_filename = argv[1]
     graphdb = GraphDatabase( graphdb_filename )
@@ -162,7 +161,7 @@ if __name__=='__main__':
         print "resources:"
         for name, resource in graphdb.resources():
             print name, resource
-        exit()
+        return
     
     vertex1 = argv[2]
     for vertex1, vertex2, edgetype in graphdb.all_outgoing( vertex1 ):
@@ -171,3 +170,6 @@ if __name__=='__main__':
         if len(argv) == 4:
             s0 = State(1,int(argv[3]))
             print "\t"+str(edgetype.walk( s0 ))
+
+if __name__=='__main__':
+    main()
